likes  y post

+++ src/main/java/dbp/techcall/post/infrastructure/PostRepository.java
@@ -5,7 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
     List<Post> findByProfessorId(Long professor_id);
+
+
 }
